ai server code refactored
