+++ ai-server/state.py
@@ -1,0 +1,2 @@
+uploaded_df = {}
+uploaded_file_info = {}
