+++ ai-server/utility/utils.py
@@ -1,0 +1,74 @@
+import json
+import numpy as np
+import pandas as pd
+import os
+from fastapi import HTTPException
+import google.generativeai as genai
+import re
+
+class NpEncoder(json.JSONEncoder):
+    def default(self, obj):
+        if isinstance(obj, np.integer):
+            return int(obj)
+        if isinstance(obj, np.floating):
+            if np.isnan(obj) or np.isinf(obj):
+                return None  # Replace NaN and Infinity with None
+            return float(obj)
+        if isinstance(obj, np.ndarray):
+            return obj.tolist()
+        if isinstance(obj, pd.Timestamp):
+            return obj.isoformat()
+        if pd.isna(obj):
+            return None
+        return super(NpEncoder, self).default(obj)
+
+def sanitize_for_json(data):
+    if isinstance(data, dict):
+        return {k: sanitize_for_json(v) for k, v in data.items()}
+    elif isinstance(data, list):
+        return [sanitize_for_json(i) for i in data]
+    elif isinstance(data, (np.int64, np.int32)):
+        return int(data)
+    elif isinstance(data, (np.float64, np.float32)):
+        return float(data)
+    elif pd.isna(data):
+        return None
+    else:
+        return data
+
+def validate_code(code: str) -> bool:
+    dangerous_patterns = [
+        r"os\.", r"subprocess\.", r"sys\.", r"shutil\.", r"eval\(", r"exec\(",
+        r"__import__", r"open\(", r"file\(", r"importlib"
+    ]
+    for pattern in dangerous_patterns:
+        if re.search(pattern, code):
+            return False
+    try:
+        import ast
+        ast.parse(code)
+        return True
+    except SyntaxError:
+        return False
+
+def get_genai_client():
+    api_key = os.environ.get("GEMINI_API_KEY")
+    if not api_key:
+        raise HTTPException(status_code=500, detail="GEMINI_API_KEY environment variable not set")
+    try:
+        genai.configure(api_key=api_key)
+        model = genai.GenerativeModel('gemini-2.0-flash')
+        return model
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Failed to initialize Gemini client: {str(e)}")
+
+def get_genai_client_thinkingmodel():
+    api_key = os.environ.get("GEMINI_API_KEY")
+    if not api_key:
+        raise HTTPException(status_code=500, detail="GEMINI_API_KEY environment variable not set")
+    try:
+        genai.configure(api_key=api_key)
+        model = genai.GenerativeModel('gemini-2.0-flash')
+        return model
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Failed to initialize Gemini client: {str(e)}")
